+++ packages/dart_frog_gen/lib/src/build_route_configuration.dart
@@ -28,6 +28,7 @@ RouteConfiguration buildRouteConfiguration(Directory directory) {
     if (globalMiddleware != null) globalMiddleware
   ];
   final routes = <RouteFile>[];
+  final rogueRoutes = <RouteFile>[];
   final directories = _getRouteDirectories(
     directory: routesDirectory,
     routesDirectory: routesDirectory,
@@ -40,15 +41,19 @@ RouteConfiguration buildRouteConfiguration(Directory directory) {
         endpoints[endpoint]!.add(file);
       }
     },
+    onRogueRoute: rogueRoutes.add,
   );
   final publicDirectory = Directory(path.join(directory.path, 'public'));
+  final mainDartFile = File(path.join(directory.path, 'main.dart'));
   return RouteConfiguration(
     globalMiddleware: globalMiddleware,
     middleware: middleware,
     directories: directories,
     routes: routes,
+    rogueRoutes: rogueRoutes,
     endpoints: endpoints,
     serveStaticFiles: publicDirectory.existsSync(),
+    invokeCustomEntrypoint: mainDartFile.existsSync(),
   );
 }
 
@@ -58,6 +63,7 @@ List<RouteDirectory> _getRouteDirectories({
   required void Function(RouteFile route) onRoute,
   required void Function(MiddlewareFile route) onMiddleware,
   required void Function(String endpoint, RouteFile file) onEndpoint,
+  required void Function(RouteFile route) onRogueRoute,
 }) {
   final directories = <RouteDirectory>[];
   final entities = directory.listSync().sorted();
@@ -87,11 +93,13 @@ List<RouteDirectory> _getRouteDirectories({
       directory: directory,
       routesDirectory: routesDirectory,
       onRoute: onRoute,
+      onRogueRoute: onRogueRoute,
     ),
     ..._getRouteFilesForDynamicDirectories(
       directory: directory,
       routesDirectory: routesDirectory,
       onRoute: onRoute,
+      onRogueRoute: onRogueRoute,
     ),
   ];
 
@@ -123,6 +131,7 @@ List<RouteDirectory> _getRouteDirectories({
           onRoute: onRoute,
           onMiddleware: onMiddleware,
           onEndpoint: onEndpoint,
+          onRogueRoute: onRogueRoute,
         ),
       );
     }
@@ -135,6 +144,7 @@ List<RouteFile> _getRouteFilesForDynamicDirectories({
   required Directory directory,
   required Directory routesDirectory,
   required void Function(RouteFile route) onRoute,
+  required void Function(RouteFile route) onRogueRoute,
   String prefix = '',
 }) {
   final files = <RouteFile>[];
@@ -149,12 +159,14 @@ List<RouteFile> _getRouteFilesForDynamicDirectories({
       directory: dynamicDirectory,
       routesDirectory: routesDirectory,
       onRoute: onRoute,
+      onRogueRoute: onRogueRoute,
       prefix: newPrefix,
     );
     final dynamicSubset = _getRouteFilesForDynamicDirectories(
       directory: dynamicDirectory,
       routesDirectory: routesDirectory,
       onRoute: onRoute,
+      onRogueRoute: onRogueRoute,
       prefix: newPrefix,
     );
     files.addAll([...subset, ...dynamicSubset]);
@@ -166,6 +178,7 @@ List<RouteFile> _getRouteFiles({
   required Directory directory,
   required Directory routesDirectory,
   required void Function(RouteFile route) onRoute,
+  required void Function(RouteFile route) onRogueRoute,
   String prefix = '',
 }) {
   final files = <RouteFile>[];
@@ -175,22 +188,36 @@ List<RouteFile> _getRouteFiles({
       ? directorySegment
       : '/$directorySegment';
   final entities = directory.listSync().sorted();
+  final subDirectories = entities
+      .whereType<Directory>()
+      .map((directory) => path.basename(directory.path))
+      .toSet();
   entities.where((e) => e.isRoute).cast<File>().forEach((entity) {
     final filePath = path
         .relative(entity.path, from: routesDirectory.path)
         .replaceAll(r'\', '/');
-    final fileRoutePath = pathToRoute(path.join('..', 'routes', filePath))
-        .split(directoryPath)
-        .last;
-    var fileRoute = fileRoutePath.isEmpty ? '/' : fileRoutePath;
-    fileRoute = prefix + fileRoute;
-    if (!fileRoute.startsWith('/')) {
-      fileRoute = '/$fileRoute';
-    }
-    if (fileRoute != '/' && fileRoute.endsWith('/')) {
-      fileRoute = fileRoute.substring(0, fileRoute.length - 1);
+
+    String getFileRoute() {
+      final routePath = pathToRoute(path.join('..', 'routes', filePath));
+      final index = routePath.indexOf(directoryPath);
+      final fileRoutePath = index == -1
+          ? routePath
+          : routePath.substring(index + directoryPath.length);
+
+      var fileRoute = fileRoutePath.isEmpty ? '/' : fileRoutePath;
+      fileRoute = prefix + fileRoute;
+
+      if (!fileRoute.startsWith('/')) {
+        fileRoute = '/$fileRoute';
+      }
+      if (fileRoute != '/' && fileRoute.endsWith('/')) {
+        fileRoute = fileRoute.substring(0, fileRoute.length - 1);
+      }
+
+      return fileRoute;
     }
 
+    final fileRoute = getFileRoute();
     final relativeFilePath = path.join('..', 'routes', filePath);
     final route = RouteFile(
       name: filePath.toAlias(),
@@ -199,6 +226,15 @@ List<RouteFile> _getRouteFiles({
     );
     onRoute(route);
     files.add(route);
+
+    final fileBasename = path.basenameWithoutExtension(filePath);
+    final conflictingIndexFile = File(
+      path.join(directory.path, fileBasename, 'index.dart'),
+    );
+    final isRogueRoute = subDirectories.contains(fileBasename) &&
+        !conflictingIndexFile.existsSync();
+
+    if (isRogueRoute) onRogueRoute(route);
   });
   return files;
 }
@@ -251,8 +287,13 @@ class RouteConfiguration {
     required this.directories,
     required this.routes,
     required this.endpoints,
+    required this.rogueRoutes,
     this.serveStaticFiles = false,
+    this.invokeCustomEntrypoint = false,
   });
+
+  /// Whether to invoke a custom entrypoint script (`main.dart`).
+  final bool invokeCustomEntrypoint;
 
   /// Whether to serve static files. Defaults to false.
   final bool serveStaticFiles;
@@ -273,6 +314,34 @@ class RouteConfiguration {
 
   /// A map of all endpoint paths to resolved route files.
   final Map<String, List<RouteFile>> endpoints;
+
+  /// List of all rogue routes.
+  ///
+  /// A route is considered rogue when it is defined outside
+  /// of an existing subdirectory with the same name.
+  ///
+  /// For example:
+  ///
+  /// ```
+  /// ├── routes
+  /// │   ├── foo
+  /// │   │   └── example.dart
+  /// │   ├── foo.dart
+  /// ```
+  ///
+  /// In the above scenario, `foo.dart` is rogue because it is defined
+  /// outside of the existing `foo` directory.
+  ///
+  /// Instead, `foo.dart` should be renamed to `index.dart` and placed within
+  /// the `foo` directory like:
+  ///
+  /// ```
+  /// ├── routes
+  /// │   ├── foo
+  /// │   │   ├── example.dart
+  /// │   │   └── index.dart
+  /// ```
+  final List<RouteFile> rogueRoutes;
 }
 
 /// {@template route_directory}
@@ -326,7 +395,7 @@ class RouteDirectory {
 }
 
 /// {@template route_file}
-/// A class containing metadata regarding a route directory.
+/// A class containing metadata regarding a route file.
 /// {@endtemplate}
 class RouteFile {
   /// {@macro route_file}
@@ -336,7 +405,7 @@ class RouteFile {
     required this.route,
   });
 
-  /// The alias for the current directory.
+  /// The alias for the current file.
   final String name;
 
   /// The import path for the current instance.