+++ packages/dart_frog/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.1.0
+
+- chore: stable 0.1.0 release
+
+# 0.0.1-dev.12
+
+- feat: expose `HttpConnectionInfo` on `Request`
+- chore: upgrade to very_good_analysis v3.0.1
+
 # 0.0.1-dev.11
 
 - fix: Request/Response `headers` is of type `Map<String, String>`